Move all configuration components into configuration folder
