+++ Source/SelfService/Web/integrations/connection/configuration/MetadataPublisherCredentials.tsx
